2025.01.02.1709/project/backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2==2.9.9
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0